// src/lib/houston/types.ts
import { Prefix } from './prefix/prefix';

/*
 *
 * Houston options object
 *
 * */

export interface Options {
  format: Format;
  logColors: Record<LogLevel, Color>;
  prefix: Prefix;
  logLevelDisplay: LogLevelDisplay;
  [key: string]: any;
}

/*
 *
 * with optional fields
 *
 * */
export interface Config {
  format?: Format;
  logColors?: Record<LogLevel, Color>;
  prefix?: Prefix;
  logLevelDisplay?: LogLevelDisplay;
  [key: string]: any;
}

/*
 *
 * Houston transport interface to implement
 *
 * */
export interface ITransport {
  level: Array<LogLevel>;
  options: Options | undefined;

  /*
   *
   * Applying the default options from Houston.ts, if custom options aren't applied already
   *
   * @param options  The default options
   *
   * */
  applyDefaultOptions(options: Options): void;

  /*
   *
   * Logging the message
   *
   * @param level  The level of the log, 0 by default
   * @param message  The log message
   *
   * */
  log(level: LogLevel, message: string): void;
}

/*
 *
 * Levels of the logs
 *
 * */
export enum LogLevel {
  Info = 'Info',
  Success = 'Success',
  Warning = 'Warning',
  Error = 'Error',
}

/*
 *
 * How to display the log level
 *
 * */
export enum LogLevelDisplay {
  Text = 'text',
  Icon = 'icon',
  Hidden = 'hidden',
}

/*
 *
 * Log level icons
 *
 * */
export enum LogLevelIcon {
  Info = '🛈',
  Success = '🗸',
  Warning = '⚠',
  Error = '✘',
}

/*
 *
 * Log format
 *
 * */
export enum Format {
  json,
  text,
}

/*
 *
 * Time format for TimePrefix
 *
 * */
export enum TimeFormat {
  American = 'american',
  European = 'european',
}

/*
 *
 * Helper enum to use colors with console.log
 *
 * */
export enum Color {
  Red = '\x1b[31m',
  Green = '\x1b[32m',
  Blue = '\x1b[34m',
  Yellow = '\x1b[33m',
  Purple = '\x1b[35m',
  Cyan = '\x1b[36m',
  Black = '\x1b[30m',
  White = '\x1b[37m',
}

/*
 *
 * Helper enum to use different styles with console.log
 *
 * */
export enum Style {
  Reset = '\x1b[0m',
  Bright = '\x1b[1m',
  Dim = '\x1b[2m',
  Underline = '\x1b[4m',
  Blink = '\x1b[5m',
  Reverse = '\x1b[7m',
  Hidden = '\x1b[8m',
}

/*
 *
 * Helper enum to use different background colors with console.log
 *
 * */
export enum Background {
  Red = '\x1b[41m',
  Green = '\x1b[42m',
  Blue = '\x1b[44m',
  Yellow = '\x1b[43m',
  Purple = '\x1b[45m',
  Cyan = '\x1b[46m',
  Black = '\x1b[40m',
  White = '\x1b[47m',
}